google_address_validation/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
pydantic==2.6.4
python-dotenv==1.0.1